Add class design
minor changes

+++ Double.h
@@ -47,7 +47,7 @@ Double operator/(const Double& lhv, const Double& rhv) {
     return lhv.value() / rhv.value();
 }
 Double operator%(const Double& lhv, const Double& rhv) {
-    return static_cast<long>(lhv.value()) % static_cast<long>(rhv.value());
+    return static_cast<double>(static_cast<long>(lhv.value()) % static_cast<long>(rhv.value()));
 }
 Double pow(const Double& lhv, const Double& rhv) {
     return pow(lhv.value(), rhv.value());
